# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yeungjin-Global",
        "description": "YJG API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/reset-password/verify": {
            "get": {
                "tags": [
                    "비밀번호"
                ],
                "summary": "비밀번호 초기화 코드 검증",
                "description": "메일로 받은 코드를 인증",
                "operationId": "9b8909ff14919ad027682e80fb9aa0ff",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "이메일로부터 받은 코드",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "사용자의 이메일",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
...
    }
}